Move account stuff into components, update stories
