+++ packages/dapp/package.json
@@ -15,11 +15,6 @@
     "@types/redux": "^3.6.0",
     "@types/sanitize-html": "^1.14.0",
     "@types/storejs": "^2.0.2",
-    "apollo-boost": "^0.1.16",
-    "apollo-cache-inmemory": "^1.3.12",
-    "apollo-client": "^2.4.8",
-    "apollo-link-context": "^1.0.12",
-    "apollo-link-http": "^1.5.9",
     "bignumber.js": "~5.0.0",
     "graphql": "^14.0.2",
     "graphql-tag": "^2.9.2",
