+++ packages/dapp/src/App.tsx
@@ -3,9 +3,9 @@ import { GlobalNav } from "./components/GlobalNav";
 import Main from "./components/Main";
 import { BrowserRouter as Router } from "react-router-dom";
 import { ApolloProvider } from "react-apollo";
-import { getClient } from "./helpers/apolloClient";
+import { getApolloClient } from "@joincivil/utils";
 
-const client = getClient({
+const client = getApolloClient({
   uri: "https://graphql.staging.civil.app/v1/query",
   // uri: "http://localhost:8080/v1/query", - use this when testing locally
 });
