+++ packages/utils/package.json
@@ -19,6 +19,11 @@
     "@types/chai": "^4.1.2",
     "@types/mocha": "^5.2.5",
     "@types/react": "^16.0.39",
+    "apollo-boost": "^0.1.16",
+    "apollo-cache-inmemory": "^1.3.12",
+    "apollo-client": "^2.4.8",
+    "apollo-link-context": "^1.0.12",
+    "apollo-link-http": "^1.5.9",
     "chai": "^4.1.2",
     "mocha": "^5.0.1",
     "react": "^16.2.0",
