+++ packages/utils/src/index.ts
@@ -1,9 +1,11 @@
 import "./rxjs";
+export * from "./apolloClient";
 export * from "./civilHelpers";
 export * from "./contracts";
 export * from "./crypto";
 export * from "./email";
 export * from "./errors";
 export * from "./language";
+export * from "./localStorage";
 export * from "./salt";
 export * from "./viewHelpers";
